+++ packages/sit-onyx/src/components/OnyxProgressSteps/OnyxProgressSteps.ct.tsx
@@ -33,6 +33,40 @@ test.describe("Screenshot tests (visited)", () => {
   });
 });
 
+test.describe("Screenshot tests (skeleton)", () => {
+  executeMatrixScreenshotTest({
+    name: "Progress steps (skeleton)",
+    columns: ORIENTATIONS,
+    rows: DENSITIES,
+    component: (column, row) => (
+      <OnyxProgressSteps density={row} orientation={column} steps={STEPS} skeleton />
+    ),
+  });
+});
+
+test.describe("Screenshot tests (max size)", () => {
+  executeMatrixScreenshotTest({
+    name: "Progress steps (max size)",
+    columns: ["default", "scrolled"],
+    rows: ORIENTATIONS,
+    component: (column, row) => (
+      <OnyxProgressSteps
+        orientation={row}
+        steps={STEPS}
+        style={{ maxWidth: "9rem", maxHeight: "7rem" }}
+      />
+    ),
+    hooks: {
+      beforeEach: async (component, page, column) => {
+        if (column === "scrolled")
+          await component
+            .getByRole("button", { name: STEPS.at(-1)!.label })
+            .scrollIntoViewIfNeeded();
+      },
+    },
+  });
+});
+
 test("should behave correctly", async ({ mount }) => {
   // ARRANGE
   const component = await mount(OnyxProgressSteps, {