+++ packages/sit-onyx/src/components/OnyxProgressSteps/OnyxProgressSteps.stories.ts
@@ -23,3 +23,10 @@ export const Vertical = {
     orientation: "vertical",
   },
 } satisfies Story;
+
+export const Skeleton = {
+  args: {
+    ...Default.args,
+    skeleton: true,
+  },
+} satisfies Story;