+++ packages/sit-onyx/src/components/OnyxProgressSteps/OnyxProgressSteps.vue
@@ -1,6 +1,7 @@
 <script lang="ts" setup>
 import { computed, watch } from "vue";
 import { useDensity } from "../../composables/density";
+import { SKELETON_INJECTED_SYMBOL, useSkeletonContext } from "../../composables/useSkeletonState";
 import { useVModel } from "../../composables/useVModel";
 import OnyxProgressStep from "../OnyxProgressStep/OnyxProgressStep.vue";
 import type { OnyxProgressStepProps, ProgressStepStatus } from "../OnyxProgressStep/types";
@@ -10,7 +11,7 @@ import type { OnyxProgressStepsProps } from "./types";
 const props = withDefaults(defineProps<OnyxProgressStepsProps>(), {
   orientation: "horizontal",
   modelValue: 1,
-  highestValue: 1,
+  skeleton: SKELETON_INJECTED_SYMBOL,
 });
 
 const emit = defineEmits<{
@@ -21,10 +22,11 @@ const emit = defineEmits<{
   /**
    * Emitted when the highest visited step/value changes.
    */
-  "update:highestValue": [value: number];
+  "update:highestValue": [value?: number];
 }>();
 
 const { densityClass } = useDensity(props);
+const skeleton = useSkeletonContext(props);
 
 const highestValue = useVModel({
   props,
@@ -58,6 +60,7 @@ const mappedSteps = computed(() => {
       value,
       disabled,
       status,
+      skeleton: skeleton.value,
       // allow user to override step properties so he has full control, therefore we add "...step" at the very end
       ...step,
     };
@@ -94,6 +97,9 @@ const mappedSteps = computed(() => {
     display: inline-flex;
     align-items: center;
     gap: var(--onyx-density-sm);
+    max-width: 100%;
+    max-height: 100%;
+    overflow: auto;
 
     &--vertical {
       flex-direction: column;