+++ packages/vitepress-theme/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sit-onyx/vitepress-theme
 
+## 1.0.0-beta.247
+
+### Patch Changes
+
+- Updated dependencies [f838364]
+  - sit-onyx@1.0.0-beta.246
+
 ## 1.0.0-beta.246
 
 ### Patch Changes